clean up imports + move config.py into lineapy
otherwise the import path of config.py will still be ambiguous.
