+++ lineapy/data/graph.py
@@ -1,4 +1,4 @@
-from typing import List, Dict, Optional, Any, cast
+from typing import cast
 
 import networkx as nx
