+++ lineapy/db/base.py
@@ -3,11 +3,11 @@ from dataclasses import dataclass
 from enum import Enum
 from typing import List
 
+from lineapy import config
 from lineapy.data.graph import Graph
 from lineapy.data.types import DataSourceNode, Node
 from lineapy.data.types import LineaID, SessionContext
 from lineapy.db.asset_manager.base import DataAssetManager
-import config
 
 
 class DatabaseOption(Enum):
