+++ lineapy/utils.py
@@ -1,4 +1,3 @@
-from tokenize import Triple
 from typing import Any, Optional
 
 IS_DEBUG = True
